Fix sign-conversion/narrowing warnings in core delay and reverb headers
Makes the existing implicit int/size_t conversions in
ParallelPlainDelay and FdnTankSpiced explicit (same computed values,
just no silent promotions), reorders BlockProcVibrato's constructor
init list to match declaration order, and switches SmoothingParameter
to the std::equal_to<float> idiom already used elsewhere in the
codebase instead of a bare ==. Verified warning-free via a clean
rebuild of all 8 examples plus the full test suite.

+++ src/includes/Delays/ParallelPlainDelay.h
@@ -22,18 +22,20 @@ class ParallelPlainDelay
     void setSize(const size_t index, const size_t newSize) noexcept
     {
         assert(index < m_read[0].size());
-        m_currentDelayWidth[index] = std::min(newSize, MAXSIZE - 1);
-        m_read[0][index] = m_head + MAXSIZE - m_currentDelayWidth[index];
-        while (m_read[0][index] >= m_buffer[index].size())
+        m_currentDelayWidth[index] = static_cast<int>(std::min(newSize, MAXSIZE - 1));
+        m_read[0][index] = static_cast<int>(m_head + MAXSIZE - static_cast<size_t>(m_currentDelayWidth[index]));
+        while (static_cast<size_t>(m_read[0][index]) >= m_buffer[index].size())
         {
-            m_read[0][index] -= m_buffer[index].size();
+            m_read[0][index] -= static_cast<int>(m_buffer[index].size());
         }
         for (size_t headIdx = 1; headIdx <= m_deltaTaps.size(); ++headIdx)
         {
-            m_read[headIdx][index] = m_head + MAXSIZE - m_currentDelayWidth[index] + m_deltaTaps[headIdx - 1][index];
-            while (m_read[headIdx][index] >= m_buffer[index].size())
+            m_read[headIdx][index] =
+                static_cast<int>(m_head + MAXSIZE - static_cast<size_t>(m_currentDelayWidth[index]) +
+                                 static_cast<size_t>(m_deltaTaps[headIdx - 1][index]));
+            while (static_cast<size_t>(m_read[headIdx][index]) >= m_buffer[index].size())
             {
-                m_read[headIdx][index] -= m_buffer[index].size();
+                m_read[headIdx][index] -= static_cast<int>(m_buffer[index].size());
             }
         }
     }
@@ -42,10 +44,11 @@ class ParallelPlainDelay
     {
         assert(headIdx > 0 && headIdx < m_read.size() && index < m_read[headIdx].size());
         m_deltaTaps[headIdx - 1][index] = newDistance;
-        m_read[headIdx][index] = m_head + MAXSIZE - m_currentDelayWidth[index] + m_deltaTaps[headIdx - 1][index];
-        while (m_read[headIdx][index] >= m_buffer[index].size())
+        m_read[headIdx][index] = static_cast<int>(m_head + MAXSIZE - static_cast<size_t>(m_currentDelayWidth[index]) +
+                                                  static_cast<size_t>(m_deltaTaps[headIdx - 1][index]));
+        while (static_cast<size_t>(m_read[headIdx][index]) >= m_buffer[index].size())
         {
-            m_read[headIdx][index] -= m_buffer[index].size();
+            m_read[headIdx][index] -= static_cast<int>(m_buffer[index].size());
         }
     }
 
@@ -70,7 +73,7 @@ class ParallelPlainDelay
         {
             for (size_t i = 0; i < BlockSize; ++i)
             {
-                target[c][i] = m_buffer[c][m_read[0][c] + i];
+                target[c][i] = m_buffer[c][static_cast<size_t>(m_read[0][c]) + i];
             }
         }
         m_head += BlockSize;
@@ -81,7 +84,7 @@ class ParallelPlainDelay
         for (size_t c = 0; c < CHANNELS; ++c)
         {
             m_read[0][c] += BlockSize;
-            if (m_read[0][c] >= MAXSIZE)
+            if (static_cast<size_t>(m_read[0][c]) >= MAXSIZE)
             {
                 m_read[0][c] -= MAXSIZE;
             }
@@ -95,13 +98,13 @@ class ParallelPlainDelay
         {
             for (size_t i = 0; i < BlockSize; ++i)
             {
-                target[c][i] = m_buffer[c][m_read[idx][c] + i];
+                target[c][i] = m_buffer[c][static_cast<size_t>(m_read[idx][c]) + i];
             }
         }
         for (size_t c = 0; c < CHANNELS; ++c)
         {
             m_read[idx][c] += BlockSize;
-            if (m_read[idx][c] >= MAXSIZE)
+            if (static_cast<size_t>(m_read[idx][c]) >= MAXSIZE)
             {
                 m_read[idx][c] -= MAXSIZE;
             }
